requirements.txt
streamlit
datasets
langchain
langchain-google-genai
langchain-chroma
langchain-huggingface
langchain-text-splitters
sentence-transformers
python-dotenv